Get rid of boto3 and move things around
