+++ dockers/llm.vdb.service/Dockerfile
@@ -40,7 +40,6 @@ RUN pip3 install --no-cache-dir \
 
 RUN pip3 install --no-cache-dir \
     "python-magic" \
-    "boto3" \
     "pydantic_settings" \
     "s3fs"
