+++ dockers/llm.vdb.service/requirements.txt
@@ -1,4 +1,3 @@
-boto3
 click
 faiss-cpu
 langchain_community
